+++ documentation/src/theme/BlogListPage/index.js
@@ -1,7 +1,6 @@
 import React from "react";
 import clsx from "clsx";
 import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
-import Link from "@docusaurus/Link";
 
 import {
     PageMetadata,
@@ -12,6 +11,7 @@ import BlogLayout from "@theme/BlogLayout";
 import BlogListPaginator from "@theme/BlogListPaginator";
 import SearchMetadata from "@theme/SearchMetadata";
 import BlogPostItems from "@theme/BlogPostItems";
+import TagsList from "@theme/TagsList";
 
 import { FeaturedBlogPostItems } from "../../components/blog";
 
@@ -53,6 +53,7 @@ function BlogListPageContent(props) {
                 </Link>
             )} */}
             <br />
+            <TagsList tags={props.tags} />
             <br />
             <br />