+++ examples/auth-mantine/src/components/header/index.tsx
@@ -34,6 +34,9 @@ export const Header: React.FC = () => {
             px="sm"
             sx={{
                 borderBottom: `1px solid ${borderColor}`,
+                position: `sticky`,
+                top: 0,
+                zIndex: 1,
             }}
         >
             <Group