+++ examples/i18n-nextjs/next-i18next.config.js
@@ -1,6 +1,9 @@
+const path = require("path");
+
 module.exports = {
     i18n: {
         defaultLocale: "en",
         locales: ["en", "de"],
+        localePath: path.resolve("./public/locales"),
     },
 };