+++ examples/monorepo-with-lerna/apps/my-refine-app/src/pages/blog-posts/create.tsx
@@ -1,0 +1,6 @@
+import { IResourceComponentsProps } from "@refinedev/core";
+import { AntdCreateInferencer } from "@refinedev/inferencer/antd";
+
+export const BlogPostCreate: React.FC<IResourceComponentsProps> = () => {
+    return <AntdCreateInferencer />;
+};