+++ packages/mantine/src/hooks/form/useModalForm/index.ts
@@ -18,12 +18,22 @@ import { UseFormInput } from "@mantine/form/lib/types";
 import React from "react";
 
 export type UseModalFormReturnType<
-    TData extends BaseRecord = BaseRecord,
+    TQueryFnData extends BaseRecord = BaseRecord,
     TError extends HttpError = HttpError,
     TVariables = Record<string, unknown>,
     TTransformed = TVariables,
-    TSelectData extends BaseRecord = TData,
-> = UseFormReturnType<TData, TError, TVariables, TTransformed, TSelectData> & {
+    TData extends BaseRecord = TQueryFnData,
+    TResponse extends BaseRecord = TData,
+    TResponseError extends HttpError = TError,
+> = UseFormReturnType<
+    TQueryFnData,
+    TError,
+    TVariables,
+    TTransformed,
+    TData,
+    TResponse,
+    TResponseError
+> & {
     modal: {
         submit: (
             values: ReturnType<
@@ -43,12 +53,22 @@ export type UseModalFormReturnType<
 };
 
 export type UseModalFormProps<
-    TData extends BaseRecord = BaseRecord,
+    TQueryFnData extends BaseRecord = BaseRecord,
     TError extends HttpError = HttpError,
     TVariables = Record<string, unknown>,
     TTransformed = TVariables,
-    TSelectData extends BaseRecord = TData,
-> = UseFormProps<TData, TError, TVariables, TTransformed, TSelectData> & {
+    TData extends BaseRecord = TQueryFnData,
+    TResponse extends BaseRecord = TData,
+    TResponseError extends HttpError = TError,
+> = UseFormProps<
+    TQueryFnData,
+    TError,
+    TVariables,
+    TTransformed,
+    TData,
+    TResponse,
+    TResponseError
+> & {
     modalProps?: {
         defaultVisible?: boolean;
         autoSubmitClose?: boolean;
@@ -57,28 +77,34 @@ export type UseModalFormProps<
 } & FormWithSyncWithLocationParams;
 
 export const useModalForm = <
-    TData extends BaseRecord = BaseRecord,
+    TQueryFnData extends BaseRecord = BaseRecord,
     TError extends HttpError = HttpError,
     TVariables = Record<string, unknown>,
     TTransformed = TVariables,
-    TSelectData extends BaseRecord = TData,
+    TData extends BaseRecord = TQueryFnData,
+    TResponse extends BaseRecord = TData,
+    TResponseError extends HttpError = TError,
 >({
     modalProps,
     refineCoreProps,
     syncWithLocation,
     ...rest
 }: UseModalFormProps<
-    TData,
+    TQueryFnData,
     TError,
     TVariables,
     TTransformed,
-    TSelectData
+    TData,
+    TResponse,
+    TResponseError
 > = {}): UseModalFormReturnType<
-    TData,
+    TQueryFnData,
     TError,
     TVariables,
     TTransformed,
-    TSelectData
+    TData,
+    TResponse,
+    TResponseError
 > => {
     const initiallySynced = React.useRef(false);
 
@@ -110,11 +136,13 @@ export const useModalForm = <
             : undefined;
 
     const useMantineFormResult = useForm<
-        TData,
+        TQueryFnData,
         TError,
         TVariables,
         TTransformed,
-        TSelectData
+        TData,
+        TResponse,
+        TResponseError
     >({
         refineCoreProps,
         ...rest,